requirements.txt
aiogram==3.10.0
requests==2.32.3
colorama==0.4.6
deep-translator==1.11.4
duckduckgo_search==6.3.3
google==3.0.0
groq==0.9.0
environs==14.2.0
tavily-python==0.5.0
lxml==5.3.0
deep-translator==1.11.4
detectlanguage==1.5.0
PyPDF2==3.0.1
pymupdf==1.24.10
google-generativeai==0.8.3
python-docx==1.1.2
pillow==10.4.0
aiohttp==3.9.5
nest-asyncio==1.6.0
youtube-transcript-api==0.6.2
pylatexenc==2.10
prettytable==3.12.0
e2b_code_interpreter
PyMovieDb==0.0.9
lxml_html_clean==0.4.1
gradio_client==1.7.0
